requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyMuPDF==1.23.7
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
SQLAlchemy==2.0.23
aiofiles==23.2.1
weasyprint==60.1
python-whois==0.8.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.2
email-validator==2.1.0.post1 